requirements.txt
# RQSM-Engine Production Dependencies
# Python 3.10+

# Web Framework
fastapi==0.109.0
uvicorn[standard]==0.27.0
pydantic==2.5.3
pydantic-settings==2.1.0

# Document Processing
pdfplumber==0.10.3
PyPDF2==3.0.1
python-docx==1.1.0
python-multipart==0.0.6

# NLP & Embeddings
sentence-transformers==2.3.1
torch==2.1.2
numpy==1.26.3
nltk==3.8.1
spacy==3.7.2

# LLM Integration (Multiple Providers)
openai==1.10.0  # Optional - only if using OpenAI
tiktoken==0.5.2
huggingface-hub==0.20.3  # For Hugging Face Inference API
google-generativeai==0.3.2  # For Google Gemini (free tier)
requests==2.31.0  # For Ollama and general HTTP

# Database
sqlalchemy==2.0.25
alembic==1.13.1

# Utilities
python-dotenv==1.0.0
loguru==0.7.2
pyyaml==6.0.1
